Remove fontawesome from covmap
This map is embedded in https://www.unocha.org/our-work/humanitarian-financing, where we discovered the 'social' buttons are going blank as they use a different version of fontawesome, which gets overwritten by the version in the map.

I had a quick look and don't *think* I see fontawesome is used in the map, so the simple answer is to remove it. Can you confirm that this is the case?

Thanks!